backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
google-generativeai==0.3.2
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2
pillow==10.2.0
numpy==1.26.4
python-magic==0.4.27
bcrypt==4.1.2
python-magic-bin==0.4.14; sys_platform == 'win32' 